refactor(database): extract DatabasePlayersMixin to database_players.py
